[#134027897] Adds anticipated departure date
Given that operations for a room that's a occupied by a "stayover" and a
room that's had it's reservation departed, it's useful to know in advance
which is happening.

Therefore the anticipated departure of the occupying reservation should be
stored along with the reference to the occupying reservation. This gets
set when the reservation occupies a room, and is maintained by the FRP
amendment rabbit event.