Adds field on room occupancy changed notification
Default-false boolean "occupancy_prematerialized_on_reservation". The
default case is that occupying a room will trigger this rabbit message,
which is picked up by booking to materialize occupancy state on the
reservation.

The non-default is for cases where we're occupying a room as part of a check-in, the check-in
transaction will materialize the occupancy state onto the reservation, in
order to avoid a race condition between the occupancy materialization and
the check-in transaction.

There are no grpc changes here because the public API should not expose
this option to not materialize state onto the reservation, and should be
a private interface within the HOLMS Application ecosystem.