[#13612473] Creates batch reporting manifest pattern
Reports are compiled into a single batch document by the server, which
recieves a manifest of the batch and its parameters from the client. This
is designed to be easily modified in most cases without causing breaking
changes.

This is a first draft of the interface. There's a lot of surface area to
these reports and some reports may jump to different manifests as I work
on incorporating them into the batch printing mechanism.