[#138683563] Adds booking channel tag
Creates a booking channel tag. This tag will contain the 3 forms of
identifying the reservation with the booking channel as well as the name
of the booking channel.

Created an "IsPermanent" field on reservation tags. This field denotes
tags that can be established in genesis, but cannot be removed by
amendment. For the forseeable future, I don't see a case for supporting
amending channel information on a reservation. Having this "IsPermament"
tag will allow the HOLMS system from Model to Client to enforce that
permanent tags are permanent.

Added channel identification information onto mail reservation DTOs. The
plan is to store the two primary booking ids (channel manager, and source
channel ids) as top-level fields on the reservation. This is for queries
where we need to look up reservations by one of these IDs.

Added the OTASyncSvc app service. This will allow us to hook the
OTASyncSvc up to the scheduler, or some other runner if we so desire.