[#141436737] Adds void reservation state
This is a tiny, but impactful proposed change.

Void state is similar to canceled. A void reservation cannot be checked
in/checked out, or cancelled. It does not contribute to supply claims, or
display on reports. It is a "This formerly-open reservation does not
exist." Void reservations are hidden from all but GetByID queries.

The purpose of this is that in the case of an OpTA booking with multiple
room stays, one room stay may commit, but another will fail. In this case
we cannot confirm to the channel manager the successful booking of the
reservation, so we remove any reservations that were created. Void is
distinct from cancelled because cancellations follow certain business
processes, and continue to impact the system. The rollback is done because
it is highly likely than a user-implemented resolution to the failed
booking would create a duplicate reservation.