Refine batch into "closed batch" and "open batch"
Open batches behave totally differently than closed ones, but it took us a
while to figure this out. When requesting the state of the open batch, we
now get signaling of the batch's status, we now get explicit batch state
signaling, including "open", "not open", or "retrieval error". I tested
and this works with HPS.

[#147801337]