[#149452341] Adds outstanding room keys to res summary
The number of outstanding room keys for a reservation gets added to the
reservation summary. This number is tracked by events, but stored as a
queriable field on the model.

The value of this is that PMC wants on its departure report a count of the number
of keys they expect a guest to have.