[#149938230] Fixes bug with == operator
EntityDTO had a bug where if the indicator of a dto was null, calling dto
== null would return true. This was happening because the null object was
getting cast up as the dto's class, and then the equality comparer would
compare dto.indicator == null.

This is fixed by a small change in the equality comparer that uses an XOR
statement to handle references to null before comparing the indicator.

Added tests as well to cover this behavior