[#153839987] Adds GetFolioSummaries RPC
Folio summary is the folio equivalent to Reservation Summary. It's a
stripped-down representation of the folio limited to only top-level fields
of the folio (no event-sourced data) This is so we have a version of the
folio that can be efficiently queried in bulk.

Added the summary to folio state to help prevent duplication